hanggout-real-time streaming/requirements.txt
firebase-admin
kafka-python
python-dotenv